+++ lib/assets/core/javascripts/cartodb3/components/modals/add-analysis/add-analysis-view.js
@@ -14,10 +14,6 @@ var DataServicesApiCheck = require('../../../editor/layers/layer-content-views/a
 module.exports = CoreView.extend({
   className: 'Dialog-content Dialog-content--expanded',
 
-  events: {
-    'click .js-add': '_onAddAnalysis'
-  },
-
   initialize: function (opts) {
     if (!opts.modalModel) throw new Error('modalModel is required');
     if (!opts.configModel) throw new Error('configModel is required');
