+++ lib/assets/core/javascripts/cartodb3/editor.js
@@ -63,7 +63,8 @@ var DataServicesApiCheck = require('./editor/layers/layer-content-views/analyses
 var AppNotifications = require('./app-notifications');
 var TipsyTooltipView = require('./components/tipsy-tooltip-view');
 var Router = require('./routes/router');
-var RouterHelpers = require('./routes/helpers');
+var handleAnalysesRoute = require('./routes/handle-analyses-route');
+var handleModalsRoute = require('./routes/handle-modals-route');
 
 // JSON data passed from entry point (editor/visualizations/show.html):
 var vizJSON = window.vizJSON;
@@ -382,7 +383,17 @@ var settingsCollection = new Backbone.Collection(SettingsOptions(overlaysCollect
 
 Router.init({
   modals: modals,
-  editorModel: editorModel
+  editorModel: editorModel,
+  routeChangeAction: function (routeModel) {
+    var currentRoute = routeModel.get('currentRoute');
+
+    if (!currentRoute) return;
+
+    handleModalsRoute(currentRoute, modals);
+    handleAnalysesRoute(currentRoute);
+
+    editorModel.set('edition', false);
+  }
 });
 
 var rootLocation = (/(\/builder\/[0-z\.\-]+\/?)/).exec(location.pathname)[1];
