+++ lib/assets/core/javascripts/cartodb3/routes/handle-analyses-route.js
@@ -1,0 +1,10 @@
+var AnalysesService = require('../editor/layers/layer-content-views/analyses/analyses-service');
+
+module.exports = function (currentRoute) {
+  var routeName = currentRoute[0];
+  var analysisNode = currentRoute[2];
+
+  if (routeName === 'layer_analyses' && !analysisNode) {
+    AnalysesService.clearNotAppliedAnalysis();
+  }
+};
