+++ lib/assets/core/javascripts/cartodb3/routes/handle-modals-route.js
@@ -1,0 +1,7 @@
+module.exports = function (currentRoute, modals) {
+  var routeName = currentRoute[0];
+
+  if (routeName !== 'modal') {
+    modals.destroy();
+  }
+};
