+++ lib/assets/core/javascripts/cartodb3/routes/router.js
@@ -1,8 +1,7 @@
 var Backbone = require('backbone');
 var _ = require('underscore');
-var populateRoute = require('./populate-route');
 var checkAndBuildOpts = require('../helpers/required-opts');
-var RouterHelpers = require('./helpers');
+var populateRoute = require('./populate-route');
 
 var ROUTE_LAYERS = 'layers';
 var ROUTE_WIDGETS = 'widgets';
@@ -22,8 +21,7 @@ var ROUTE_ADD_POLYGON = BASE_LAYER_ROUTE + '/add/polygon';
 var EDIT_FEATURE = BASE_LAYER_ROUTE + '/edit/:featureId';
 
 var REQUIRED_OPTS = [
-  'modals',
-  'editorModel'
+  'routeChangeAction'
 ];
 
 var ROUTES = [
@@ -47,7 +45,7 @@ var ROUTES = [
   ['basemap', ROUTE_BASEMAP]
 ];
 
-var Router = (function () {
+module.exports = (function () {
   var initialized = false;
   var routeModel = new Backbone.Model();
   // keep last route before a modal, so we can replace /modal with it
@@ -89,17 +87,7 @@ var Router = (function () {
     },
 
     _initBinds: function () {
-      this.getRouteModel().on('change:currentRoute', function (routeModel) {
-        var route = routeModel.get('currentRoute');
-        if (!route) return;
-        if (route[0] !== 'modal') {
-          this._modals.destroy();
-        }
-
-        RouterHelpers.handleAnalysesRoute(routeModel);
-
-        this._editorModel.set('edition', false);
-      }, this);
+      this.getRouteModel().on('change:currentRoute', this._routeChangeAction);
     },
 
     getRouter: function () {
@@ -245,5 +233,3 @@ var Router = (function () {
     }
   };
 })();
-
-module.exports = Router;
