+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cartodb-ui",
-  "version": "4.11.11",
+  "version": "4.11.11.13364.01",
   "description": "CARTO UI frontend",
   "repository": {
     "type": "git",
